0929 댓글 삭제 기능 구현 삽질만 하다가 끝남
며칠째 삭제 기능구현이 도저히 안되서 삽질중..

+++ src/main/java/com/hong/fitness/controller/FitnessBoardController.java
@@ -50,7 +50,7 @@ public class FitnessBoardController {
 	
 	// 게시판 글보기
 	@GetMapping("/view.do")
-	public String view(@RequestParam("no") long no, Model model, HttpSession session, FitnessBoardVO vo, PageObject pageObject) throws Exception {
+	public String view(@RequestParam("no") int no, Model model, HttpSession session, FitnessBoardVO vo, PageObject pageObject) throws Exception {
 		
 		log.info("fitness board 글보기 no : " + no);
 		
@@ -122,4 +122,8 @@ public class FitnessBoardController {
 		
 		return "redirect:list.do?perPageNum=" + perPageNum;
 	}
+	
+	// 댓글 삭제 get
+	//public String replyDelete
+	
 }
